refactor: move Runtime from wcore to runtime crate
wcore becomes pure types/traits — Hook, Conversation, Storage,
Agent, Config types stay. The execution engine (Runtime struct with
send_to, stream_to, agent registry, conversation management) moves
to crates/runtime/src/engine.rs where it belongs alongside Env.
