+++ crates/node/src/node/builder.rs
@@ -9,14 +9,14 @@ use crate::{
 use anyhow::Result;
 use crabllm_core::Provider;
 use crabllm_provider::{ProviderRegistry, RemoteProvider};
-use runtime::{Env, host::Host, memory::Memory};
+use runtime::{Env, Runtime, host::Host, memory::Memory};
 use std::{
     collections::{BTreeMap, HashMap},
     path::{Path, PathBuf},
     sync::Arc,
 };
 use tokio::sync::{Mutex, RwLock, broadcast};
-use wcore::{AgentConfig, Runtime, ToolRequest, model::Model, repos::Storage};
+use wcore::{AgentConfig, ToolRequest, model::Model, repos::Storage};
 use wcore::{ResolvedManifest, resolve_manifests};
 
 pub type DefaultProvider = crate::provider::Retrying<ProviderRegistry<RemoteProvider>>;
