+++ crates/node/src/node/mod.rs
@@ -13,13 +13,13 @@ use crate::{
 };
 use anyhow::Result;
 use crabllm_core::Provider;
-use runtime::{Env, host::Host};
+use runtime::{Env, Runtime, host::Host};
 use std::{
     path::{Path, PathBuf},
     sync::Arc,
 };
 use tokio::sync::{Mutex, RwLock, broadcast, mpsc, oneshot};
-use wcore::{Runtime, model::Model};
+use wcore::model::Model;
 
 pub(crate) mod builder;
 pub mod event;
