+++ crates/runtime/Cargo.toml
@@ -12,11 +12,15 @@ keywords.workspace = true
 
 [dependencies]
 wcore.workspace = true
+crabllm-core.workspace = true
 
 # crates.io
 anyhow.workspace = true
-reqwest.workspace = true
+async-stream.workspace = true
 chrono.workspace = true
+futures-core.workspace = true
+futures-util.workspace = true
+reqwest.workspace = true
 schemars.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -26,5 +30,7 @@ toml.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
+futures-util.workspace = true
 tempfile.workspace = true
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 wcore = { workspace = true, features = ["test-utils"] }
