+++ crates/runtime/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod ask_user;
 pub mod config;
+mod engine;
 pub mod env;
 pub mod host;
 pub mod memory;
@@ -8,6 +9,7 @@ pub mod skill;
 pub mod task;
 
 pub use config::{MemoryConfig, SystemConfig, TasksConfig};
+pub use engine::Runtime;
 pub use env::Env;
 pub use host::{Host, NoHost};
 pub use memory::Memory;
